+++ NEWS.md
@@ -1,6 +1,7 @@
 # mFD (development version)
 * Fix bug:
   * Correct the sp.plot() fct when plotting one asb (pb with vertices aes)
+  * Correct the sp.plot() fct when several sp same place but only one as vert
 
 # mFD 1.0.5
 * Fix bug: